Woopsie! Did a oopsie doos!
Fixing this as I accidentally went off the master branch and I didn't know how to do this without a normal merge so... Yeah Hi this is all the stuff!

Added a new hub cap to walls.
Added damage mechanics to thirst and hunger (using existing code that was unused)
Added new sprites to the Medical Analyzer with thanks from Scuhf_
Added a new dungeon type (W.I.P Presets and Pack)
Added a new song to the lobby / round end

fml honestly